requirements/common.txt
# SPDX-FileCopyrightText: Magenta ApS
#
# SPDX-License-Identifier: MPL-2.0

structlog

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-aiohttp-client
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-fastapi

os2mo-http-trigger-protocol==0.0.3
